move components to components file and display correct answer
