+++ week-sonia-asmahan/src/components/ResetGame.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import Flags from "./Flags";
 
 export default function ResetGame(props) {
   if (props.lives !== 0) {
@@ -12,7 +11,7 @@ export default function ResetGame(props) {
     );
   } else if (props.lives === 0) {
     return (
-      <div class="quiz-container">
+      <div className="quiz-container">
         <p>Game Over</p>
         <button className="btn" onClick={() => props.restartGame()}>
           Play again
